// script/types.ts
export type RawNavNode = Record<string, RawNavNode[] | string>;

/** Unparsed YML configurations. */
export interface RawConfig {
    /** Navigation data. */
    nav: RawNavNode[];

    /** Category icons. */
    icon: Record<string, string>;
}

/** Parsed nav node. */
export interface NavNode {
    /** Nav title */
    title: string;

    /** Nav name, e.g. `cs`, `math` */
    name: string;

    /** Link to the article */
    link: string;

    /** Children nodes */
    children: NavNode[];
}

/** Parsed configurations. */
export type Config = Omit<RawConfig, "nav"> & { nav: NavNode[] };